Add parse_cubfile.c and more
