+++ srcs/init_utils/parse_cubfile_utils/set_direction.c
